backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# AI / Embeddings
openai==1.30.1
sentence-transformers==3.0.0
pinecone-client==3.2.2

# Document processing
pypdf2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.1
nltk==3.8.1
langchain-text-splitters==0.2.0

# HTTP / async
httpx==0.27.0
aiofiles==23.2.1

# Caching
redis==5.0.4

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
loguru==0.7.2
slowapi==0.1.9
